feat: enable k/v cache quantisation
