+++ llm/memory_test.go
@@ -128,88 +128,3 @@ func TestEstimateGPULayers(t *testing.T) {
 		})
 	}
 }
-
-func TestEstimateKvCacheSize(t *testing.T) {
-	const (
-		baseSize = 0x10000000 // Base size for standard model with 1024 context using f32
-		baseCtx  = uint64(1024)
-	)
-
-	var (
-		tokenizer = []string{" "}
-		scores    = []float32{0}
-		tokenType = []int32{0}
-	)
-
-	// Helper to create a test model with given architecture
-	createTestModel := func(arch string, poolingType string) *GGML {
-		f, err := os.CreateTemp(t.TempDir(), "dummy")
-		require.NoError(t, err)
-		defer f.Close()
-
-		tensor := Tensor{
-			Name:     "output.weight",
-			Kind:     uint32(0),
-			Offset:   uint64(0),
-			Shape:    []uint64{1, 1, 1, 1},
-			WriterTo: bytes.NewReader(make([]byte, 32)),
-		}
-
-		kv := KV{
-			"general.architecture":            arch,
-			arch + ".context_length":          uint32(32),
-			arch + ".embedding_length":        uint32(4096),
-			arch + ".block_count":             uint32(32),
-			arch + ".attention.head_count":    uint32(32),
-			arch + ".attention.head_count_kv": uint32(32),
-			"tokenizer.ggml.tokens":           tokenizer,
-			"tokenizer.ggml.scores":           scores,
-			"tokenizer.ggml.token_type":       tokenType,
-		}
-
-		if poolingType != "" {
-			kv[arch+".pooling_type"] = poolingType
-		}
-
-		require.NoError(t, WriteGGUF(f, kv, []Tensor{tensor}))
-
-		ggml, err := LoadModel(f.Name(), 0)
-		require.NoError(t, err)
-		return ggml
-	}
-
-	tests := []struct {
-		name        string
-		ctx         uint64
-		cacheType   string
-		isEmbedding bool
-		want        uint64
-	}{
-		{"f16 standard", baseCtx, "f16", false, baseSize * 2},     // f16 uses 2 bytes
-		{"q8_0 standard", baseCtx, "q8_0", false, baseSize},       // q8_0 uses 1 byte
-		{"q4_0 standard", baseCtx, "q4_0", false, baseSize / 2},   // q4_0 uses 0.5 bytes
-		{"empty type", baseCtx, "", false, baseSize * 2},          // defaults to f16
-		{"unknown type", baseCtx, "unknown", false, baseSize * 2}, // defaults to f16
-		{"f16 embedding", baseCtx, "f16", true, baseSize * 2},     // embedding models still use same memory
-		{"q8_0 embedding", baseCtx, "q8_0", true, baseSize},       // q8_0 still uses 1 byte
-		{"q4_0 embedding", baseCtx, "q4_0", true, baseSize / 2},   // q4_0 still uses 0.5 bytes
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			arch := "llama"
-			poolingType := ""
-			if tt.isEmbedding {
-				poolingType = "mean"
-			}
-
-			ggml := createTestModel(arch, poolingType)
-
-			got := estimateKvCacheSize(tt.cacheType, tt.ctx, ggml)
-
-			assert.Equal(t, tt.want, got,
-				fmt.Sprintf("ctx=%d requested_cache=%s embedding=%v",
-					tt.ctx, tt.cacheType, tt.isEmbedding))
-		})
-	}
-}
